requirements.txt
torch==1.5
torchvision==0.6
numpy==1.18.1
imageio==2.8.0
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.4.1
Pillow==7.0.0
opencv-python==3.4.1.15
h5py==2.10.0
pandas==0.25.3
Cython==0.29.14